Don't allow FBT to instrument DDB, which makes debugging FBT quite hard.
Don't allow FBT to instrument memcpy().  This is a bit lazy, but avoids
several panic variations I've encountered.  I need to figure out if it
is because memcpy() is being substituted by the compiler, or whether there
is an undesirable explicit call to memcpy() somewhere in DTrace or its
dependencies.